# Conflicts:
#	src/main/java/com/catalyst/collector/controllers/CollectionsController.java
#	src/main/java/com/catalyst/collector/daos/CollectionsDao.java
#	src/main/java/com/catalyst/collector/daos/hibernate/CollectionsDaoHibernate.java
#	src/main/java/com/catalyst/collector/entities/Collectible.java
#	src/main/java/com/catalyst/collector/services/CollectionsService.java
#	src/main/java/com/catalyst/collector/services/impl/CollectionsServiceImpl.java

+++ src/main/java/com/catalyst/collector/daos/hibernate/CollectionsDaoHibernate.java
@@ -30,25 +30,51 @@ public class CollectionsDaoHibernate implements CollectionsDao {
 
 
 
-	public void addAge(Age age){
-		em.persist(age);
+	public boolean addAge(Age age){
+		String ageString = age.getAge();
+		if(ageString.length() < 256 && !ageString.matches(".*\\d.*")) {
+			em.persist(age);
+			return true;
+		}
+		return false;
+
 	}
 
 	public ArrayList<Age> getAgeTypes(){
 		return (ArrayList<Age>) em.createQuery("SELECT t FROM Age t", Age.class).getResultList();
 	}
 
-	public void updateAge(Age age){
-		em.merge(age);
+	public boolean updateAge(Age age){
+		String ageString = age.getAge();
+		if(ageString.length() < 256 && !ageString.matches(".*\\d.*")) {
+			em.merge(age);
+			return true;
+		}
+		return false;
 	}
 
-	public void deleteAge(Integer id){
-		Age age = em
-				.createQuery("SELECT e FROM Age e WHERE e.id = :id", Age.class)
-				.setParameter("id", id)
-				.getSingleResult();
-		em.remove(age);
+	public boolean deleteAge(Integer id){
+		if(id > 0) {
+			Age age = em
+					.createQuery("SELECT e FROM Age e WHERE e.id = :id", Age.class)
+					.setParameter("id", id)
+					.getSingleResult();
+			em.remove(age);
+			return true;
+		}
+		return false;
 	}
+
+	@Override
+	public void addCollectible(Collectible collectible) {
+		em.persist(collectible);
+	}
+
+	@Override
+	public void updateCollectible(Collectible collectible) {
+		em.merge(collectible);
+	}
+
 
     @Override
     public ArrayList<Condition> getAllConditions() {
@@ -71,17 +97,7 @@ public class CollectionsDaoHibernate implements CollectionsDao {
         em.remove(condition);
     }
 
-	@Override
-	public void addCollectible(Collectible collectible) {
-		em.persist(collectible);
-	}
-
-	@Override
-	public void updateCollectible(Collectible collectible) {
-		em.merge(collectible);
-	}
-
-	@Override
+    @Override
 	public ArrayList<Category> getCategory() {
 		return (ArrayList<Category>)em
 				.createQuery("SELECT c FROM Category c", Category.class)
@@ -185,7 +201,7 @@ public class CollectionsDaoHibernate implements CollectionsDao {
 	}
 	@Override
 	public ArrayList<Collectible> getCollectibles() {
-		return (ArrayList<Collectible>) em.createQuery("Select c from Collectible c", Collectible.class).getResultList();
+		return (ArrayList<Collectible>) em.createQuery("Select c FROM Collectible c", Collectible.class).getResultList();
 	}
 
 	@Override
